refactor: くるもに合わせたディレクトリ構成に変更
