+++ lib/splash.dart
@@ -1,0 +1,2 @@
+/// くるもにあったのでいったん作成
+/// └splash.dart　 アプリメイン
